:zap: Allow @ConstructorBinding to be optional
This commit makes @ConstructorBinding optional for a type that has a single parameterized constructor. An @Autowired annotation on any of the constructors indicates that the type should not be constructor bound.  Since @ConstructorBinding is now deduced for a single parameterized constructor, the annotation is no longer needed at the type level.