:zap: Introduce fluent API for searches in MergedAnnotations
Prior to this commit, searching for merged annotations on an AnnotatedElement in the MergedAnnotations model was only supported via various overloaded from(...) factory methods. In addition, it was not possible to provide a custom AnnotationFilter without providing an instance of RepeatableContainers.  This commit introduces a fluent API for searches in MergedAnnotations to address these issues and improve the programming model for users of MergedAnnotations.  To begin a search, invoke MergedAnnotations.search(SearchStrategy) with the desired search strategy. Optional configuration can then be provided via one of the with(...) methods. To perform a search, invoke from(AnnotatedElement), supplying the element from which to begin the search -- for example, a Class or a Method.  For example, the following performs a search on MyClass within the entire type hierarchy of that class while ignoring repeatable annotations.  MergedAnnotations mergedAnnotations = MergedAnnotations.search(SearchStrategy.TYPE_HIERARCHY) .withRepeatableContainers(RepeatableContainers.none()) .from(MyClass.class);  To reuse search configuration to perform the same type of search on multiple elements, you can save the Search instance as demonstrated in the following example.  Search search = MergedAnnotations.search(SearchStrategy.TYPE_HIERARCHY) .withRepeatableContainers(RepeatableContainers.none());  MergedAnnotations mergedAnnotations = search.from(MyClass.class); // do something with the MergedAnnotations for MyClass mergedAnnotations = search.from(AnotherClass.class); // do something with the MergedAnnotations for AnotherClass  In addition, this fluent search API paves the way for introducing support for a predicate that controls the search on enclosing classes (gh-28207) and subsequently for completely removing the TYPE_HIERARCHY_AND_ENCLOSING_CLASSES search strategy (gh-28080).  Closes gh-28208