:zap: Provide a public API for Application hooks
Create a new public Application Hook API based on the existing
`ApplicationStartupListener` interface.

The previous package-private `ApplicationHook` class has been
replaced with a public `ApplicationHook` interface which acts as
a factory that can create additional `ApplicationStartupListener`
instances to hook in.

The boolean result from the previous `preRefresh` method has been
replaced with an `AbandonedRunException` which can be thrown from
the `ApplicationStartupListener`.