Refine constructor detection logic when binding to existing values
Update `DefaultBindConstructorProvider` so that deduced constructors
are not used if there is an existing value.

Prior to this commit, constructor detection logic was not compatible, given
a class of the following form:

    @ConfigurationProperties(prefix = "example")
    public class ExampleProperties {

        @NestedConfigurationProperty
        private final NestedProperty nested = new NestedProperty(
                "Default", "default");

        public NestedProperty getNested() {
            return nested;
        }

    }

If `NestedProperty` has a single constructor with arguments, constructor
binding would be used.

The updated code now only uses constructor injection if an explicit
`@ConstructorBinding` annotation is present, or if there is no existing
value.