:sparkles: Add SSL bundle support to today-core module
Add classes to support SSL bundles which can be used to apply SSL
settings in a centralized way. An `SslBundle` can be registered with
an `SslBundleRegistry` and obtained from an `SslBundles` instance. The
`DefaultSslBundleRegistry` provides a default in-memory implementation.

Different client libraries often configure SSL in slightly different
ways. To accommodate this, the `SslBundle` provides a layered approach
of obtaining SSL information:

    - `getStores` provides access to the key store and trust stores
      as well as any required key store password.

    - `getManagers` provides access to the `KeyManagerFactory`,
      `TrustManagerFactory` as well as the `KeyManger` and
      `TrustManager` arrays that they create.

    - `createSslContext` provides a convenient way to obtain a new
      `SSLContext` instance.

In addition, the `SslBundle` also provides details about the key being
used, the protocol to use and any options that should be applied to the
SSL engine.